// react-module/src/openapi/models/non-erp-production-order-params.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface NonErpProductionOrderParams
 */
export interface NonErpProductionOrderParams {
    /**
     * 
     * @type {number}
     * @memberof NonErpProductionOrderParams
     */
    nonErpOrderId: any;
    /**
     * 
     * @type {number}
     * @memberof NonErpProductionOrderParams
     */
    amount: any;
}