// react-module/src/openapi/models/production-order-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ProductionOrderExtendedDto
 */
export interface ProductionOrderExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof ProductionOrderExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {number}
     * @memberof ProductionOrderExtendedDto
     */
    orderElementPosition?: any;
    /**
     * 
     * @type {number}
     * @memberof ProductionOrderExtendedDto
     */
    amount?: any;
    /**
     * 
     * @type {LayerShapeExtendedDto}
     * @memberof ProductionOrderExtendedDto
     */
    layerShape?: any;
    /**
     * 
     * @type {Array&lt;AtomicOperationExtendedDto&gt;}
     * @memberof ProductionOrderExtendedDto
     */
    operationsList?: any;
    /**
     * 
     * @type {Array&lt;WithoutOrder&gt;}
     * @memberof ProductionOrderExtendedDto
     */
    completedProducts?: any;
    /**
     * 
     * @type {AtomicProductDefinitionBasicDto}
     * @memberof ProductionOrderExtendedDto
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {LocalErpOrderDto}
     * @memberof ProductionOrderExtendedDto
     */
    erpOrder?: any;
}